[review] 입력 검증 부분 switch제거, 각 타입별로 체크할수 있도록 함
