+++ Signup/Signup/Sources/Common/Verification/Type/EmailValidate.swift
@@ -1,0 +1,17 @@
+//
+//  EmailVerification.swift
+//  Signup
+//
+//  Created by seongha shin on 2022/03/31.
+//
+
+import Foundation
+
+enum EmailValidate: BaseValidate {
+    static func verification(text: String) -> ValidateResultType {
+        if validatePredicate(text, format: "^[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,20}$") {
+            return .success
+        }
+        return .errorEmail
+    }
+}
