+++ Signup/Signup/Sources/Common/Verification/Verification.swift
@@ -1,0 +1,14 @@
+//
+//  Verification.swift
+//  Signup
+//
+//  Created by seongha shin on 2022/03/31.
+//
+
+import Foundation
+
+class Verification<T: BaseValidate> {
+    func check(text: String) -> ValidateResultType {
+        T.verification(text: text)
+    }
+}
